backend/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Werkzeug==3.0.4
authlib==1.3.2
openai==1.51.2
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg[binary]==3.2.2
stripe==10.12.0
